edited parsechunk.py to fix the whitespace problem
