+++ test.py~
@@ -1,0 +1,9 @@
+import parsechunk.py
+import df.py
+
+copy_lines('/projectnb/scv/atime/projectnb_econdept.list',' /projectnb/rcs-intern/alana/objects/parsed.csv')
+
+df = load_data('/projectnb/rcs-intern/alana/objects/parsed.csv')
+
+df.head()
+
